+++ lib/callable_tree/node/internal.rb
@@ -3,7 +3,10 @@
 module CallableTree
   module Node
     module Internal
+      extend ::Forwardable
       include Node
+
+      def_delegators :child_nodes, :[], :at
 
       def children
         # TODO: Change to return a new array instance.