+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Change `CallableTree::Node#match?` to accept inputs to the node as variable length arguments.
 - Change `CallableTree::Node#call` to accept inputs to the node as variable length arguments.
 - Change `CallableTree::Node#terminate?` to accept inputs to the node as variable length arguments, after the `output` argument.
+- Add `CallableTree::Node::Internal#[]` to return the child node using `index`.
 
 ## [0.2.3] - 2021-11-07